Make the suite portable to 32-bit, Windows, and a nightly compiler bug
Three tests assumed Int is Int64 or drew from a range that overflows Int32;
they now name Int64 explicitly. The owned-bytes test rewrites a file the
table was parsed from; Windows keeps a mapped file locked until the mapping
is finalized, so the test collects first and retries. Julia 1.14.0-DEV.2945
miscompiles `f(x) === nothing ? a : b` when f returns Union{Nothing, Type}
(a package-free reproduction is recorded in the review scratchpad), which
breaks the transposed reader and the batch schema pass on nightly only; the
nightly CI job is marked as allowed to fail until that is fixed upstream.

+++ test/fuzz.jl
@@ -251,7 +251,7 @@ end
         end
     end
     @testset "late promotion, masked adoption, and retained scalars" begin
-        values = [string(rand(rng, 10^14:10^15)) for _ in 1:300]
+        values = [string(rand(rng, Int64(10)^14:Int64(10)^15)) for _ in 1:300]   # 32-bit safe
         values[151] = "long text " * "λ"^40
         values[end] = "escaped \""^20
         io = IOBuffer()
